Desktop/Doc2Tasks/requirements.txt
__future__ 
pickle
os.path
googleapiclient.discovery
google_auth_oauthlib.flow
google.auth.transport.requests
csv
math
json
re
argparse
